Move NameValidator from package org.molgenis.data.meta to org.molgenis.data.validation.meta
