+++ molgenis-core-ui/src/main/java/org/molgenis/data/importer/wizard/OptionsWizardPage.java
@@ -8,8 +8,8 @@ import org.molgenis.data.RepositoryCollection;
 import org.molgenis.data.importer.EntitiesValidationReport;
 import org.molgenis.data.importer.ImportService;
 import org.molgenis.data.importer.ImportServiceFactory;
-import org.molgenis.data.meta.NameValidator;
 import org.molgenis.data.meta.model.Package;
+import org.molgenis.data.validation.meta.NameValidator;
 import org.molgenis.ui.wizard.AbstractWizardPage;
 import org.molgenis.ui.wizard.Wizard;
 import org.molgenis.util.FileExtensionUtils;
