+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/meta/NameValidator.java
@@ -1,4 +1,4 @@
-package org.molgenis.data.meta;
+package org.molgenis.data.validation.meta;
 
 import com.google.common.collect.ImmutableSet;
 import org.molgenis.data.MolgenisDataException;
@@ -10,6 +10,10 @@ import java.util.Set;
  */
 public class NameValidator
 {
+	private NameValidator()
+	{
+	}
+
 	// some words are reserved for the RestAPI and default packages/entities/attributes, etc.
 	public static final Set<String> KEYWORDS = ImmutableSet.of("login", "logout", "csv", "base", "exist", "meta");
