+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/meta/PackageValidator.java
@@ -2,7 +2,6 @@ package org.molgenis.data.validation.meta;
 
 import org.molgenis.data.MolgenisDataException;
 import org.molgenis.data.meta.MetaUtils;
-import org.molgenis.data.meta.NameValidator;
 import org.molgenis.data.meta.model.Package;
 import org.molgenis.data.meta.system.SystemPackageRegistry;
 import org.springframework.stereotype.Component;
