+++ molgenis-one-click-importer/src/main/java/org/molgenis/oneclickimporter/service/OneClickImporterNamingService.java
@@ -1,6 +1,6 @@
 package org.molgenis.oneclickimporter.service;
 
-import org.molgenis.data.meta.NameValidator;
+import org.molgenis.data.validation.meta.NameValidator;
 
 /**
  * Service that offers methods to:
